[FEAT/#1607] print refactor
